refactor: package 구조 수정  - domain, view layer 분리
