+++ src/main/java/racingcar/domain/Moving.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.domain;
 
 public class Moving {
     public boolean move(int generatedNumber) {
