+++ src/test/java/racingcar/MovingTest.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
+import racingcar.domain.Moving;
 
 @DisplayName("움직임 테스트")
 public class MovingTest {
