+++ src/test/java/racingcar/RandomNumberGeneratorTest.java
@@ -3,6 +3,7 @@ package racingcar;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import racingcar.domain.RandomNumberGenerator;
 
 @DisplayName("랜덤 숫자 생성 테스트")
 public class RandomNumberGeneratorTest {
